requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.3.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
flake8==7.3.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.5.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-telegram-bot==22.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
